requirements.txt
Flask  
datetime 
uuid
PyJWT
psycopg